requirements.txt
# Multi-Agent System for Cross-Domain Genie Queries
# Python 3.11+ required

# Core Agent Framework
langgraph-supervisor==0.0.30
langchain-core>=0.1.0
langchain>=0.1.0

# Databricks Integration
mlflow[databricks]>=2.9.0
databricks-langchain>=0.1.0
databricks-agents>=0.1.0
databricks-vectorsearch>=0.22
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Configuration & Validation
pydantic>=2.0.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
urllib3>=2.0.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
ipykernel>=6.25.0
notebook>=7.0.0

# Optional: For local development
jupyter>=1.0.0
ipython>=8.12.0

